Address second review round on #688
- CONFLATED_PARQUET.md: 'byte-for-byte' -> 'point-for-point'.
- src/geometry/wkb_type.rs: link geoparquet_name() to the
  geometry_types section of the GeoParquet 2.0-rc.1 spec (no anchor
  available there yet).
- Move GeometryTally from pipeline::conflate::writer into a new
  src/geometry/geometry_tally.rs, re-exported as crate::geometry::
  GeometryTally, and generalize its doc comment away from
  conflated.parquet's two specific columns -- filed #690 for reusing
  it in places::writer once the ATP phase actually preserves real
  (non-point) geometry, which is the prerequisite for that logging
  to say anything but "Point" every time.

Verified: cargo test/clippy/fmt clean.
