+++ src/geometry/mod.rs
@@ -41,6 +41,7 @@ use geo::{
 };
 
 mod geometry_builder;
+mod geometry_tally;
 #[cfg(test)]
 mod grid_tests;
 mod line_stitcher;
@@ -49,6 +50,7 @@ mod polygon_union;
 mod wkb_type;
 
 pub use geometry_builder::{GeometryBuilder, PolygonFill};
+pub use geometry_tally::GeometryTally;
 pub use line_stitcher::LineStitcher;
 pub use polygon_assembler::PolygonAssembler;
 pub use polygon_union::PolygonUnion;
